+++ README.md
@@ -68,7 +68,7 @@ A meta-repository to organize my public projects.
 
 ## Usage
 
-The [table above](#current) and [plots below](#plots) are generated by the script [`git-repos.py`](git-repos.py).
+The [table above](#current) and [plots below](#plots) are generated by the snakemake workflow in the [`Snakefile`](Snakefile).
 
 ### Setup
 
@@ -90,25 +90,51 @@ The [table above](#current) and [plots below](#plots) are generated by the scrip
     conda env create -n git-repos -f environment.yaml
     conda activate git-repos
     ```
+    
+### Config
+
+Generate an [access token](https://help.github.com/en/articles/creating-a-personal-access-token-for-the-command-line)
+ and paste it into a plain text file.
+ 
+ Update [`config.yml`](config.yml) with the path to your token file:
+ 
+```
+token_filename: path/to/token
+```
+
+Don't share your token with anyone!
 
 ### Run it
 
 Get your repos & make plots with:
 
 ```
-python git-repos.py --username your_username
+snakemake
 ```
 
-It will then prompt you for your GitHub password, collect the information from GitHub, and generate the table in [`README.md`](README.md) under the [`Projects`](#projects) subheading.
-[Plots](##plots) of programming language statistics are written to [`figures/`](figures/).
+#### Alternate workflow
+
+If you don't want to use snakemake, you can run the Python & R scripts in [`code/`](code).
+
+```
+python code/GitHubLangStats.py --token path_to_token
+```
+
+It will then prompt you for your GitHub password, collect the information from GitHub, write a csv file to [`data/`]
+(data), and generate the table in [`README.md`](README.md) under the [`Projects`](#projects) subheading.
 Only repositories which you own or which you have contributed to are included.
 
-Alternatively, you can supply an [access token](https://help.github.com/en/articles/creating-a-personal-access-token-for-the-command-line) instead of your username & password.
-Paste the token into a plain text file and use the `--token` option:
+Alternatively, you can supply your username. It will then prompt you for your password:
+```
+python code/GitHubLangStats.py --username your_username
+```
 
+To make the plots, run the R script:
 ```
-python git-repos.py --token path_to_token
+R -e 'code/plot_language_stats.R
 ```
+
+[Plots](##plots) of programming language statistics are written to [`figures/`](figures/).
 
 #### Private repos
 
@@ -117,13 +143,12 @@ However, they are always included in statistics for plots.
 If you wish to include them, use the flag `--include_private`:
 
 ```
-python git-repos.py --token path_to_token --include_private
+python code/GitHubLangStats.py --token path_to_token --include_private
 ```
 or
 ```
-python git-repos.py --username your_username --include_private
+python code/GitHubLangStats.py --username your_username --include_private
 ```
-
 
 ## Plots
